pyproject.toml
[project]
name = "w40k-explorer"
version = "0.1.0"
description = "Warhammer 40k RAG application with scraped wiki content and Streamlit chat interface"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core scraping and data processing
    "beautifulsoup4>=4.12.0",
    "aiohttp>=3.9.0",
    "tqdm>=4.66.0",
    # RAG and AI
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "openai>=1.12.0",
    "qdrant-client>=1.7.0",
    "tiktoken>=0.6.0",
    # UI
    "streamlit>=1.31.0",
    "streamlit-chat>=0.1.1",
    # Utilities
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.24",
    "lxml>=6.0.0",
    "types-beautifulsoup4>=4.12.0.20250516",
    "pydantic-settings>=2.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "black>=24.0.0",
]
enhancements = [
    "cohere>=4.47.0",
    "networkx>=3.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/w40k"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "W", "F", "I", "B", "C4", "ARG", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
